backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2==2.9.9
django-phonenumber-field==7.3.0
phonenumbers==8.13.0
PyJWT==2.11.0
djangorestframework-simplejwt==5.5.1
python-telegram-bot==20.3
google-auth==2.38.0
Pillow==11.1.0
drf-writable-nested==0.7.0
requests==2.32.3
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
whitenoise==6.7.0
